python-flask/requirements.txt
connexion >= 2.7.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
pykeybasebot
pyyaml
